fix link & move to component
